backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.3

# Wagtail CMS
wagtail==7.2.1
wagtail-modeladmin==2.0.0

# Authentication
django-allauth==0.61.1
dj-rest-auth==5.0.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Payments
stripe==8.0.0

# Print-on-Demand Integration
requests==2.31.0

# Environment & Configuration
python-decouple==3.8

# Image Processing
Pillow==10.2.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Utilities
python-dateutil==2.8.2

# Calendar Sync
icalendar==5.0.11
